+++ src/test/java/baseEntities/BaseCucumberTest.java
@@ -3,5 +3,5 @@ package baseEntities;
 import org.openqa.selenium.WebDriver;
 
 public class BaseCucumberTest {
-    public static WebDriver driver;
+   public static WebDriver driver;
 }
